+++ packages/go_router_builder/example/lib/readme_excerpts.g.dart
@@ -18,140 +18,173 @@ List<RouteBase> get $appRoutes => [
 
 RouteBase get $homeRoute => GoRouteData.$route(
       path: '/',
-      factory: $HomeRouteExtension._fromState,
+      factory: _$HomeRoute._fromState,
       routes: [
         GoRouteData.$route(
           path: 'family/:fid',
-          factory: $FamilyRouteExtension._fromState,
+          factory: _$FamilyRoute._fromState,
         ),
       ],
     );
 
-extension $HomeRouteExtension on HomeRoute {
+mixin _$HomeRoute on GoRouteData {
   static HomeRoute _fromState(GoRouterState state) => const HomeRoute();
 
+  @override
   String get location => GoRouteData.$location(
         '/',
       );
 
+  @override
   void go(BuildContext context) => context.go(location);
 
+  @override
   Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
+  @override
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
 
+  @override
   void replace(BuildContext context) => context.replace(location);
 }
 
-extension $FamilyRouteExtension on FamilyRoute {
+mixin _$FamilyRoute on GoRouteData {
   static FamilyRoute _fromState(GoRouterState state) => FamilyRoute(
         fid: state.pathParameters['fid'],
       );
 
+  FamilyRoute get _self => this as FamilyRoute;
+
+  @override
   String get location => GoRouteData.$location(
-        '/family/${Uri.encodeComponent(fid ?? '')}',
+        '/family/${Uri.encodeComponent(_self.fid ?? '')}',
       );
 
+  @override
   void go(BuildContext context) => context.go(location);
 
+  @override
   Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
+  @override
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
 
+  @override
   void replace(BuildContext context) => context.replace(location);
 }
 
 RouteBase get $loginRoute => GoRouteData.$route(
       path: '/login',
-      factory: $LoginRouteExtension._fromState,
+      factory: _$LoginRoute._fromState,
     );
 
-extension $LoginRouteExtension on LoginRoute {
+mixin _$LoginRoute on GoRouteData {
   static LoginRoute _fromState(GoRouterState state) => LoginRoute(
         from: state.uri.queryParameters['from'],
       );
 
+  LoginRoute get _self => this as LoginRoute;
+
+  @override
   String get location => GoRouteData.$location(
         '/login',
         queryParams: {
-          if (from != null) 'from': from,
+          if (_self.from != null) 'from': _self.from,
         },
       );
 
+  @override
   void go(BuildContext context) => context.go(location);
 
+  @override
   Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
+  @override
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
 
+  @override
   void replace(BuildContext context) => context.replace(location);
 }
 
 RouteBase get $myRoute => GoRouteData.$route(
       path: '/my-route',
-      factory: $MyRouteExtension._fromState,
+      factory: _$MyRoute._fromState,
     );
 
-extension $MyRouteExtension on MyRoute {
+mixin _$MyRoute on GoRouteData {
   static MyRoute _fromState(GoRouterState state) => MyRoute(
         queryParameter:
             state.uri.queryParameters['query-parameter'] ?? 'defaultValue',
       );
 
+  MyRoute get _self => this as MyRoute;
+
+  @override
   String get location => GoRouteData.$location(
         '/my-route',
         queryParams: {
-          if (queryParameter != 'defaultValue')
-            'query-parameter': queryParameter,
+          if (_self.queryParameter != 'defaultValue')
+            'query-parameter': _self.queryParameter,
         },
       );
 
+  @override
   void go(BuildContext context) => context.go(location);
 
+  @override
   Future<T?> push<T>(BuildContext context) => context.push<T>(location);
 
+  @override
   void pushReplacement(BuildContext context) =>
       context.pushReplacement(location);
 
+  @override
   void replace(BuildContext context) => context.replace(location);
 }
 
 RouteBase get $personRouteWithExtra => GoRouteData.$route(
       path: '/person',
-      factory: $PersonRouteWithExtraExtension._fromState,
+      factory: _$PersonRouteWithExtra._fromState,
     );
 
-extension $PersonRouteWithExtraExtension on PersonRouteWithExtra {
+mixin _$PersonRouteWithExtra on GoRouteData {
   static PersonRouteWithExtra _fromState(GoRouterState state) =>
       PersonRouteWithExtra(
         state.extra as Person?,
       );
 
+  PersonRouteWithExtra get _self => this as PersonRouteWithExtra;
+
+  @override
   String get location => GoRouteData.$location(
         '/person',
       );
 
-  void go(BuildContext context) => context.go(location, extra: $extra);
+  @override
+  void go(BuildContext context) => context.go(location, extra: _self.$extra);
 
+  @override
   Future<T?> push<T>(BuildContext context) =>
-      context.push<T>(location, extra: $extra);
+      context.push<T>(location, extra: _self.$extra);
 
+  @override
   void pushReplacement(BuildContext context) =>
-      context.pushReplacement(location, extra: $extra);
+      context.pushReplacement(location, extra: _self.$extra);
 
+  @override
   void replace(BuildContext context) =>
-      context.replace(location, extra: $extra);
+      context.replace(location, extra: _self.$extra);
 }
 
 RouteBase get $hotdogRouteWithEverything => GoRouteData.$route(
       path: '/:ketchup',
-      factory: $HotdogRouteWithEverythingExtension._fromState,
+      factory: _$HotdogRouteWithEverything._fromState,
     );
 
-extension $HotdogRouteWithEverythingExtension on HotdogRouteWithEverything {
+mixin _$HotdogRouteWithEverything on GoRouteData {
   static HotdogRouteWithEverything _fromState(GoRouterState state) =>
       HotdogRouteWithEverything(
         _$boolConverter(state.pathParameters['ketchup']!)!,
@@ -159,23 +192,30 @@ extension $HotdogRouteWithEverythingExtension on HotdogRouteWithEverything {
         state.extra as Sauce,
       );
 
+  HotdogRouteWithEverything get _self => this as HotdogRouteWithEverything;
+
+  @override
   String get location => GoRouteData.$location(
-        '/${Uri.encodeComponent(ketchup.toString())}',
+        '/${Uri.encodeComponent(_self.ketchup.toString())}',
         queryParams: {
-          if (mustard != null) 'mustard': mustard,
+          if (_self.mustard != null) 'mustard': _self.mustard,
         },
       );
 
-  void go(BuildContext context) => context.go(location, extra: $extra);
+  @override
+  void go(BuildContext context) => context.go(location, extra: _self.$extra);
 
+  @override
   Future<T?> push<T>(BuildContext context) =>
-      context.push<T>(location, extra: $extra);
+      context.push<T>(location, extra: _self.$extra);
 
+  @override
   void pushReplacement(BuildContext context) =>
-      context.pushReplacement(location, extra: $extra);
+      context.pushReplacement(location, extra: _self.$extra);
 
+  @override
   void replace(BuildContext context) =>
-      context.replace(location, extra: $extra);
+      context.replace(location, extra: _self.$extra);
 }
 
 bool _$boolConverter(String value) {