+++ packages/go_router_builder/test_inputs/named_escaped_route.dart
@@ -4,5 +4,7 @@
 
 import 'package:go_router/go_router.dart';
 
+mixin _$NamedEscapedRoute {}
+
 @TypedGoRoute<NamedEscapedRoute>(path: '/named-route', name: r'named$Route')
-class NamedEscapedRoute extends GoRouteData {}
+class NamedEscapedRoute extends GoRouteData with _$NamedEscapedRoute {}