+++ packages/go_router/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 15.1.3
+
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+* Fixes typo in API docs.
+
 ## 15.1.2
 
 - Fixes focus request propagation from `GoRouter` to `Navigator` by properly handling the `requestFocus` parameter.